fix: routing with urls and prefix-all-locales/patterns to treat en the same
We want non-localized routes to be handled dynamically via redirects, and
 all content localized to xx to be static under the URL xx/some-page.
We expect this to work best with Netlify, and be least confusing.

The Paraglide 2.0 framework needs both prefix-all-locales as a
 routing strategy and also for us to specify all patterns so as to force /en/
 as a prefix - the default patterns don't truly honor prefix-all-locales.

Implementation in writeSettingsFile() since it has the final locale list.

Since visiting a URL now sets the locale cookie, also added a visual effect
 to the locale switcher button so folk can see when this has occurred.

Also includes a bugfix: recent LLM message translations had added commentary
 around the JSON object, which meant Paraglide couldn't parse those messages.
 I've implemented protection against this at last responsible moment (before
 compiling) but can move it earlier when we work more on LLM improvements.