Fix unlocalized logo link - use localizeHref for homepage
The logo link was hardcoded to '/' which meant Spanish users on /es/ pages
would be taken back to the English homepage when clicking the logo.

Now uses localizeHref('/') to properly localize the link, so Spanish users
stay in /es/ when clicking the logo.

Fixes part of the issues described in #448